Dev Script
I think there was a way to do this with Rust itself, but this is what I have been doing for this project for the time being.

Forgot to link the page I was looking at for the previous few commits. I got help from GPT to help explain the difference between TS's modules and Rust's modules, as my own misconceptions seemed to be holding me back to thinking in what way it should work. Rust's modules are not file-scoped, they are more explicit than that.

https://doc.rust-lang.org/book/ch07-04-bringing-paths-into-scope-with-the-use-keyword.html